Commit incl two major updates: interactive tree and new meta-data retrieval
These features; the interactive tree and new meta-data retrieval have been added to the Advanced Search page, the strain info page and the groupwise comparisons page. The interactive tree was submitted in a previous commit, several significant changes were made to allow single and multi node selection. The new meta-data retrieval preruns the meta data queries stores them in the meta table. These are set as json obj to the user when form webages are requested.